Broke new API in common_play_skill.py
Broke out CommonPlaySkill as done in the Spotify skill To simplify synching while
working through the shape of the common base class in several skills.
